refactor: favorite domain 분리
