+++ src/main/kotlin/backend/itracker/tracker/favorite/service/vo/FavoriteCount.kt
@@ -1,4 +1,4 @@
-package backend.itracker.tracker.member.service.vo
+package backend.itracker.tracker.favorite.service.vo
 
 import backend.itracker.crawl.common.ProductCategory
